Moved tree view util modules and added comments to classes
