+++ packages/foam-vscode/src/features/panels/orphans.ts
@@ -2,8 +2,8 @@ import * as vscode from 'vscode';
 import { Foam } from '../../core/model/foam';
 import { createMatcherAndDataStore } from '../../services/editor';
 import { getOrphansConfig } from '../../settings';
-import { GroupedResourcesTreeDataProvider } from '../../utils/grouped-resources-tree-data-provider';
-import { ResourceTreeItem, UriTreeItem } from '../../utils/tree-view-utils';
+import { GroupedResourcesTreeDataProvider } from './utils/grouped-resources-tree-data-provider';
+import { ResourceTreeItem, UriTreeItem } from './utils/tree-view-utils';
 import { IMatcher } from '../../core/services/datastore';
 import { FoamWorkspace } from '../../core/model/workspace';
 import { FoamGraph } from '../../core/model/graph';
