+++ packages/foam-vscode/src/features/panels/utils/tree-view-utils.ts
@@ -1,14 +1,14 @@
 import * as vscode from 'vscode';
 import { groupBy } from 'lodash';
-import { Resource } from '../core/model/note';
-import { toVsCodeUri } from './vsc-utils';
-import { Range } from '../core/model/range';
-import { URI } from '../core/model/uri';
-import { FoamWorkspace } from '../core/model/workspace';
-import { getNoteTooltip } from '../utils';
-import { isSome } from '../core/utils';
-import { getBlockFor } from '../core/services/markdown-parser';
-import { FoamGraph } from '../core/model/graph';
+import { Resource } from '../../../core/model/note';
+import { toVsCodeUri } from '../../../utils/vsc-utils';
+import { Range } from '../../../core/model/range';
+import { URI } from '../../../core/model/uri';
+import { FoamWorkspace } from '../../../core/model/workspace';
+import { getNoteTooltip } from '../../../utils';
+import { isSome } from '../../../core/utils';
+import { getBlockFor } from '../../../core/services/markdown-parser';
+import { FoamGraph } from '../../../core/model/graph';
 
 export class BaseTreeItem extends vscode.TreeItem {
   resolveTreeItem(): Promise<vscode.TreeItem> {
